Küçük sistemler daha eklendi.

+++ packages/api/src/PharmaDesk.API/Controllers/NotificationsController.cs
@@ -102,10 +102,10 @@ namespace Backend.Controllers
         }
 
 
-        private int? GetPharmacyIdFromToken()
+        private long? GetPharmacyIdFromToken()
         {
             var pharmacyIdClaim = User.FindFirst("PharmacyId")?.Value;
-            if (int.TryParse(pharmacyIdClaim, out var pharmacyId))
+            if (long.TryParse(pharmacyIdClaim, out var pharmacyId))
                 return pharmacyId;
             return null;
         }
